Get rid of munchhausen.
The check has been irrelevant for a very long time, since asofold reworked the vertical handling with vDistRel.
It was coded as a preemptive measure for potential bypasses, back when NCP's vertical check was based on jump phases counting (more than a decade ago).
The fix was also questionable at best, as it basically interfered with a vanilla feature (hitting oneself with a rod).