Attempt to fix hidden micro moves, revert order of negligible momentum checking.
When the player comes to a stop, after several micro (and split) moves, the actual stop isn't sent to the server. The next move will yield a false positive due to the player moving off with motion based on a previous move of 0.0, which the client did not send.
There doesn't seem to be a way to detect this gap in movement, thus this brute-force case.

Bonus:
* Cleanup
* Comments
* Simplify sprinting checks.