Fix several aspects of vdistrel; rename class; remove moving trace flags;
For lost ground, we know only (and always) check for the "from" position, since it is the one we're interested in, when checking if the momentum initialisation should be set to 0.0. The "to" position is irrelevant and should only be used to further confine workarounds and such.
As a result, the lost ground brute force has been removed from both predictions.
We require last y motion to be explicitly negative in such case.

The 0.2 motion ovveride was in the wrong place, so it should now be correct, however for lower versions (1.21.1-), we need to loop the space bar press.

Removed several flags from PlayerMoveData in favour of more simple flags in CombinedData, as we don't need to track more than the last state for these specific effects/status. Also, from 1.21.2 and onwards, the input data should be used to read sprinting and sneaking.

Adjusted lost ground conditions a bit (horizontal motion requirements, y acceleration condition)

Renamed InputDirection to InputState, as we now track more than just WASD impulses.

Fixed an NPE in SupportingBlockUtils.