Fix yOnGround parameter for fastbreak; revise block push handling for SurvivalFly.
The ground-seeking parameter was abnormally large, and no information or documentation was left for us about it. However, it does seem to cause false positives when breaking blocks below, hence the fix.
We now just use the normal yOnGround parameter for movement checks.

Block push is now prioritized in predictions: the vertical one is skipped altogether, and the movement is simply allowed.
For the horizontal one, it’s a bit trickier: the prediction is skipped only if the push occurred on both axes at the same time, saving some resources. If the push occurred on only one axis, the prediction still runs for both, but the result is overridden for the axis on which the push occurred.
This is inefficient performance-wise; we’ll need a refactor to allow NCP to skip predictions on a single axis only.